+++ packages/punchcard/lib/analytics/index.ts
@@ -1,4 +1,4 @@
-export * from './lake';
+export * from './data-lake';
+export * from './data-pipeline';
 export * from './period';
-export * from './pipeline';
 export * from './schema';
