+++ packages/punchcard/lib/firehose/index.ts
@@ -1,3 +1,4 @@
 export * from './collector';
 export * from './delivery-stream';
 export * from './event';
+export * from './stream';
